LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_hit_test_scroll_background.html

<!doctype html>

<meta charset=utf-8>
<title>pointer-events: auto descendant correctly targets scrolls</title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

    <style>
        .scroller {
            width: 400px;
            height: 400px;
            border: 1px solid black;
            padding: 10px;
            overflow: scroll;
            font-size: 16pt;
            background-color: rgba(255, 255, 255, 0.75);
            margin-bottom: 100px;
        }

        h1 {
            font-size: 24pt;
        }

        .contents {
            height: 400%;
        }
    </style>

<body id="body">
    <div id="background" class="scroller">
        <div class="contents">Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</div>
    </div>

    <div id="foreground" class="scroller" style="left: 100px; top: 100px; pointer-events: none; position:absolute ">
        <h1>pointer-events: none</h1>
        <p>Scrolling here should scroll the scroller behind</p>
        <div class="contents">Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</div>
    </div>
</body>

<script>
  promise_test(async (t) => {
    let scrolled = new Promise((resolve) => {
      let scrollers = [window, document.getElementById("background"), document.getElementById("foreground")];
      let onscroll = (evt) => {
        for (const scroller of scrollers) {
          scroller.removeEventListener("scroll", onscroll);
        }
        resolve(evt.target.id || "root");
      }
      for (const scroller of scrollers) {
        scroller.addEventListener("scroll", onscroll);
      }
    });
    const actions = new test_driver.Actions().scroll(200, 200, 0, 50, { duration: 50 });
    actions.send();
    assert_equals(await scrolled, "background", "Incorrect element scrolled");
  }, "pointer-events: none foreground should not prevent background to be scrolled.");
</script>
